reputation signal on correct, wrong action
